requirements.txt
streamlit
pytesseract
Pillow
scikit-learn
joblib
